refactor: update the architecture Change to feature-oriented folder structure
