+++ src/components/screens/home/home-screen.tsx
@@ -1,9 +1,9 @@
-import { useGetAllCategories } from 'api-hook/category/use-get-all-categories'
 import DefaultPublicLayout from 'components/templates/public-layout/default-public-layout'
 import { Column } from 'components/toolkit'
 import Spinner from 'components/toolkit/spinner/spinner'
-import CategoryProductsList from 'features/product/product-list/category-products-list'
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
+import { useGetAllCategories } from 'features/category/hooks/use-get-all-categories'
+import CategoryProductsList from 'features/product/components/product-list/category-products-list'
 import React from 'react'
 
 // export interface HomeScreenProps {}
@@ -12,7 +12,12 @@ const HomeScreen: React.FC = () => {
 
   return (
     <DefaultPublicLayout className="space-y-12 py-12" title="Home">
-      {isLoading && <Column className={classNames('items-center')}><Spinner color="black" size={100} />Carregando...</Column>}
+      {isLoading && (
+        <Column className={classNames('items-center')}>
+          <Spinner color="black" size={100} />
+          Carregando...
+        </Column>
+      )}
       {categories?.map((category) => (
         <CategoryProductsList
           key={category.name}
