+++ src/components/screens/products/products-screen.tsx
@@ -1,11 +1,11 @@
-import { useGetAllProducts } from 'api-hook/product/use-get-all-products'
-import CategoryLateralMenuLayout from 'components/category-lateral-menu-layout/category-lateral-menu-layout'
 import DefaultPublicLayout from 'components/templates/public-layout/default-public-layout'
 import { Column, Row } from 'components/toolkit'
 import Spinner from 'components/toolkit/spinner/spinner'
+import { classNames } from 'core/helpers/class-names'
 import { ProductsList } from 'features/category'
-import RelatedSearches from 'features/related-searches/related-searches'
-import { classNames } from 'helpers/class-names'
+import CategoryLateralMenuLayout from 'features/category/components/category-lateral-menu-layout/category-lateral-menu-layout'
+import RelatedSearches from 'features/category/components/related-searches/related-searches'
+import { useGetAllProducts } from 'features/product/hooks/use-get-all-products'
 
 const ProductsScreen: React.FC = () => {
   const { isLoading } = useGetAllProducts()
@@ -15,7 +15,12 @@ const ProductsScreen: React.FC = () => {
       <RelatedSearches />
       <Row className="justify-between mb-2">
         <CategoryLateralMenuLayout />
-        {isLoading && <Column className={classNames('items-center')}><Spinner color="black" size={100} />Carregando...</Column>}
+        {isLoading && (
+          <Column className={classNames('items-center')}>
+            <Spinner color="black" size={100} />
+            Carregando...
+          </Column>
+        )}
         <ProductsList />
       </Row>
     </DefaultPublicLayout>
