+++ src/components/templates/public-layout/default-footer/footer-component.tsx
@@ -1,4 +1,4 @@
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 import React from 'react'
 
 interface FooterComponentProps {
