+++ src/components/templates/public-layout/default-header/header-logo.tsx
@@ -1,6 +1,6 @@
 import { CheckIcon, ShoppingBagIcon } from '@heroicons/react/outline'
 import Row from 'components/toolkit/row/row'
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 import React from 'react'
 
 export interface HeaderLogoProps {
