+++ src/components/toolkit/button/button.tsx
@@ -1,5 +1,5 @@
-import { classNames } from 'helpers/class-names';
-import React from 'react';
+import { classNames } from 'core/helpers/class-names'
+import React from 'react'
 
 /**
  * @contentButton text/icon
@@ -9,26 +9,22 @@ import React from 'react';
  //  */
 
 export interface ButtonProps {
-    contentButton: React.ReactNode;
-    className: string;
-    type?: "button" | "submit" | "reset"
-    onClick?: () => void;
+  contentButton: React.ReactNode
+  className: string
+  type?: 'button' | 'submit' | 'reset'
+  onClick?: () => void
 }
 const Button: React.FC<ButtonProps> = ({
-    contentButton,
-    className,
-    type,
-    onClick,
+  contentButton,
+  className,
+  type,
+  onClick,
 }) => {
-    return (
-        <button
-            className={classNames(className)}
-            type={type}
-            onClick={onClick}
-        >
-            {contentButton}
-        </button>
-    );
-};
+  return (
+    <button className={classNames(className)} type={type} onClick={onClick}>
+      {contentButton}
+    </button>
+  )
+}
 
-export default Button;
+export default Button
