+++ src/components/toolkit/card.tsx
@@ -1,5 +1,5 @@
 import { Column, ColumnProps } from 'components/toolkit'
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 import React from 'react'
 
 export type CardProps = ColumnProps
