+++ src/components/toolkit/column/column.tsx
@@ -1,4 +1,4 @@
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 
 export type ColumnProps = React.HTMLAttributes<HTMLDivElement>
