+++ src/components/toolkit/row/row.tsx
@@ -1,4 +1,4 @@
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 
 export type RowProps = React.HTMLAttributes<HTMLDivElement>
