+++ src/components/toolkit/text/text.tsx
@@ -1,4 +1,4 @@
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 
 export interface TextProps {
   children?: React.ReactNode
