+++ src/core/types/custom-error.ts
