+++ src/features/category/components/category-lateral-menu-layout/stock-number.tsx
