+++ src/features/category/components/related-searches/index.ts
