+++ src/features/category/components/related-searches/related-searches.tsx
@@ -1,5 +1,5 @@
 import Row from 'components/toolkit/row/row'
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 import React from 'react'
 
 const RelatedSearches: React.FC = () => {
@@ -8,16 +8,20 @@ const RelatedSearches: React.FC = () => {
     'camisas masculinas',
     'body feminino',
     'cropped',
-    'conjuntos femininos'
+    'conjuntos femininos',
   ]
 
   return (
-    <Row className='py-4'>
-      <h3 className={classNames('whitespace-nowrap line-clamp-1 font-bold text-gray-700')}>
+    <Row className="py-4">
+      <h3
+        className={classNames(
+          'whitespace-nowrap line-clamp-1 font-bold text-gray-700'
+        )}
+      >
         Buscas relacionadas:
       </h3>
       {relatedList.map((relatedItem, i) => (
-        <Row className='overflow-hidden' key={relatedItem}>
+        <Row className="overflow-hidden" key={relatedItem}>
           <p
             className={classNames(
               'line-clamp-1 ml-2 truncate text-gray-500 font-normal'
