+++ src/features/product/components/product-card/product-card.tsx
@@ -1,10 +1,10 @@
 import { Card, Column, Row, Text } from 'components/toolkit'
 import Badge from 'components/toolkit/badge'
 import { BadgeVariant } from 'components/toolkit/badge/badge'
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
+import { Product } from 'features/product/types/product'
 import Link from 'next/link'
 import React from 'react'
-import { Product } from 'types/product'
 import ProductImage from './product-image'
 import ProductInstallment from './product-installment'
 import ProductPrice from './product-price'
