+++ src/features/product/components/product-card/product-image.tsx
@@ -1,5 +1,5 @@
 import { Column } from 'components/toolkit'
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 import React from 'react'
 
 export interface ProductImageProps
