+++ src/features/product/components/product-card/product-price.tsx
@@ -1,5 +1,5 @@
 import { Row } from 'components/toolkit'
-import { classNames } from 'helpers/class-names'
+import { classNames } from 'core/helpers/class-names'
 import React from 'react'
 
 export type ProductPriceSize = 'small' | 'medium' | 'large'
