+++ src/features/product/types/product.ts
