--- src/types/card-component-props.ts
@@ -1,9 +1,0 @@
-export type ProductInstallmentType = {
-  quantity: number
-  value: number
-}
-
-export type ProductImageType = {
-  src?: string
-  alt?: string
-}
